requirements.txt
certifi==2019.11.28
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
Django==3.2.4
django-allauth==0.41.0
django-bootstrap3==12.0.3
django-cors-headers==3.8.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
entrypoints==0.3
idna==2.8
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.15.2
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
markdown==3.3.4
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==21.0
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
Pygments==2.5.2
pyparsing==2.4.7
pyrsistent==0.15.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
Send2Trash==1.5.0
simplejson==3.17.3
six==1.13.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.7
webencodings==0.5.1
psycopg2==2.9.1
dj_database_url==0.5.0
